Implemented setting of preserving original images

+++ includes/class-wsi-hooks.php
@@ -128,7 +128,7 @@ class WSI_Hooks {
 		add_filter( 'wp_handle_upload', array($this, __FUNCTION__), 10, 2 );
 
 		// Important! Unlink the temp file, but measure size beforehand
-		if ( file_exists( $orig_file['file'] ) ) {
+		if ( winsite_image_optimizer()->retro->get_setting( 'should-keep-original' ) != '1' && file_exists( $orig_file['file'] ) ) {
 			$this->last_file_size = filesize( $orig_file['file'] );
 			// Goodbye.
 			unlink( $orig_file['file'] );
